+++ src/system.d
@@ -25,7 +25,7 @@ real strToDouble(string str, uint base=10)
 	real ret = 0;
 	real mantissa = 0;
 	
-	foreach(uint i;0..str.length) {
+	foreach(size_t i;0..str.length) {
 		if(str[i] == '.') {
 			mantissa = 1;
 		} else {
